+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Shader parser now handles `ColorMask 0` passes with no code block correctly and throws a parse error for other non code block passes.
 * Fix optimizer not checking all the animator controllers in the avatar descriptor for some things.
 * Fix special animation layers not getting updated animation clips.
+* Layers with only a off animation no longer block default disabled meshes from using NaNimation. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/126)
 
 ## v3.8.0
 ### Features